requirements.txt
# Phish Detector - Requirements
# Python 3.11+ recommended

# Data processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Machine Learning
scikit-learn>=1.3.0,<2.0.0
joblib>=1.3.0,<2.0.0

# API Framework
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.23.0,<1.0.0
pydantic>=2.0.0,<3.0.0

# Date parsing
python-dateutil>=2.8.0,<3.0.0

# Testing
pytest>=7.4.0,<9.0.0
pytest-cov>=4.1.0,<6.0.0

# Visualization
matplotlib>=3.8.0,<4.0.0
seaborn>=0.13.0,<1.0.0

# Optional: for development
# httpx>=0.24.0  # for testing FastAPI
# black>=23.0.0  # code formatting
# isort>=5.12.0  # import sorting
# mypy>=1.5.0    # type checking
